requirements.txt
requests
Pillow
labelme
azure-ai-documentintelligence